+++ gocd/__init__.py
@@ -1,5 +1,5 @@
 # -*- encoding: utf-8 -*-
-__version__ = '0.9.0'
+__version__ = '0.10.0'
 __all__ = ['Server']
 
 from .server import Server