+++ gocd/api/pipeline.py
@@ -1,3 +1,6 @@
+import time
+
+from gocd.api.response import Response
 from gocd.api.endpoint import Endpoint
 
 
@@ -110,7 +113,8 @@ class Pipeline(Endpoint):
 
         return self._get('/instance/{counter:d}'.format(counter=counter))
 
-    def schedule(self, variables=None, secure_variables=None, materials=None):
+    def schedule(self, variables=None, secure_variables=None, materials=None,
+                 return_new_instance=False, backoff_time=1.0):
         """Schedule a pipeline run
 
         Aliased as :meth:`run`, :meth:`schedule`, and :meth:`trigger`.
@@ -123,6 +127,13 @@ class Pipeline(Endpoint):
             have a look at the official
             `Go pipeline scheduling documentation`__ or inspect a call
             from triggering manually in the UI.
+          return_new_instance (bool): Returns a :meth:`history` compatible
+            response for the newly scheduled instance. This is primarily so
+            users easily can get the new instance number. **Note:** This is done
+            in a very naive way, it just checks that the instance number is
+            higher than before the pipeline was triggered.
+          backoff_time (float): How long between each check for
+            :arg:`return_new_instance`.
 
          .. __: http://api.go.cd/current/#scheduling-pipelines
 
@@ -137,7 +148,28 @@ class Pipeline(Endpoint):
 
         scheduling_args = dict((k, v) for k, v in scheduling_args.items() if v is not None)
 
-        return self._post('/schedule', ok_status=202, **scheduling_args)
+        # TODO: Replace this with whatever is the official way as soon as gocd#990 is fixed.
+        # https://github.com/gocd/gocd/issues/990
+        if return_new_instance:
+            last_run = self.history()['pipelines'][0]['counter']
+            response = self._post('/schedule', ok_status=202, **scheduling_args)
+            if not response:
+                return response
+
+            max_tries = 10
+            while max_tries > 0:
+                current = self.instance()
+                if current['counter'] > last_run:
+                    return current
+                else:
+                    time.sleep(backoff_time)
+                    max_tries -= 1
+
+            # I can't come up with a scenario in testing where this would happen, but it seems
+            # better than returning None.
+            return response
+        else:
+            return self._post('/schedule', ok_status=202, **scheduling_args)
     #: This is an alias for :meth:`schedule`
     run = schedule
     #: This is an alias for :meth:`schedule`